+++ sentry/src/main/java/io/sentry/protocol/User.java
@@ -52,7 +52,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    *
    * @param email the e-mail.
    */
-  public void setEmail(@Nullable String email) {
+  public void setEmail(final @Nullable String email) {
     this.email = email;
   }
 
@@ -70,7 +70,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    *
    * @param id the user id.
    */
-  public void setId(@Nullable String id) {
+  public void setId(final @Nullable String id) {
     this.id = id;
   }
 
@@ -88,7 +88,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    *
    * @param username the username.
    */
-  public void setUsername(@Nullable String username) {
+  public void setUsername(final @Nullable String username) {
     this.username = username;
   }
 
@@ -106,7 +106,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    *
    * @param ipAddress the IP address of the user.
    */
-  public void setIpAddress(@Nullable String ipAddress) {
+  public void setIpAddress(final @Nullable String ipAddress) {
     this.ipAddress = ipAddress;
   }
 
@@ -124,8 +124,12 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    *
    * @param other the other user related data..
    */
-  public void setOthers(@Nullable Map<String, String> other) {
-    this.other = new ConcurrentHashMap<>(other);
+  public void setOthers(final @Nullable Map<String, String> other) {
+    if (other != null) {
+      this.other = new ConcurrentHashMap<>(other);
+    } else {
+      this.other = null;
+    }
   }
 
   /**
@@ -135,7 +139,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    */
   @ApiStatus.Internal
   @Override
-  public void acceptUnknownProperties(Map<String, Object> unknown) {
+  public void acceptUnknownProperties(final @NotNull Map<String, Object> unknown) {
     this.unknown = new ConcurrentHashMap<>(unknown);
   }
 
@@ -145,6 +149,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    * @return the unknown map
    */
   @TestOnly
+  @Nullable
   Map<String, Object> getUnknown() {
     return unknown;
   }